+++ lib/enums/sound_volume_view_state.dart
@@ -1,0 +1,10 @@
+
+enum SoundVolumeViewState {
+  notInstalled,
+  uninstalled,
+  errorToUninstall,
+  installed,
+  errorToInstall,
+  platformNotSupported,
+  errorToSaveFileEncoding
+}