+++ lib/sound_volume_view_library.dart
@@ -3,3 +3,5 @@ library sound_volume_view;
 export 'models/devices.dart';
 export 'core/sound_volume_view.dart';
 export 'enums/default_type.dart';
+export 'enums/sound_volume_view_state.dart';
+export 'exception/sound_volume_view_exception.dart';